Add function to reduce CFF n without modifying memory (just change variable value) and fix types related to a CFF's n to be long long, not int

+++ src/cff_internals.h
@@ -11,7 +11,7 @@ struct cff
     int d;
     int t;
     long long n;
-    int stride_bits; // helper for faster memory access
+    long long stride_bits; // helper for faster memory access
     unsigned char *matrix;
 };
